Main Locations page
I made the Main Locations page and adapter.
This is untested, simply due to my lack of knowledge regarding I/O,
however the ground is set and simply requires a few extra lines when I
know what I'm testing, where to test for (regarding changing icons) and
what the result is.
The areas that require development marked with // REQ ATTN